refactor: split fallow hotspot helpers
