+++ browser/server.mjs
@@ -4,12 +4,10 @@ import path from "node:path"
 import { fileURLToPath } from "node:url"
 
 import { buildMaintenancePlan } from "../lib/maintenance-scheduler.mjs"
-import {
-  clampInteger,
-  normalizeRepository,
-  parseJsonArray,
-  parseJsonObject,
-} from "../lib/data-utils.mjs"
+import { clampInteger } from "../lib/numeric-utils.mjs"
+import { parseJsonArray } from "../lib/json-array-utils.mjs"
+import { parseJsonObject } from "../lib/json-object-utils.mjs"
+import { normalizeRepository } from "../lib/repository-utils.mjs"
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
 const STATIC_ROOT = __dirname
