+++ lib/capsule-assembler.mjs
@@ -5,7 +5,7 @@ import {
   isStyleAddressingMemory,
 } from "./style-addressing.mjs";
 import { QUERY_ALIASES } from "./query-normalizer.mjs";
-import { normalizeText } from "./text-normalizer.mjs";
+import { normalizeText } from "./prompt-text-normalizer.mjs";
 import { estimateTokens } from "./token-estimator.mjs";
 
 const STOP_WORDS = new Set([
