+++ lib/content-normalizer.mjs
@@ -1,0 +1,3 @@
+export function normalizeText(value) {
+  return String(value || "").replace(/\s+/g, " ").trim();
+}
