+++ lib/db-episode-scoring.mjs
@@ -7,7 +7,7 @@ import {
   QUERY_ALIASES,
   tokenizeText,
 } from "./query-normalizer.mjs";
-import { normalizeText } from "./text-normalizer.mjs";
+import { normalizeText } from "./session-text-normalizer.mjs";
 import { readHybridRetrievalEnabled } from "./rollout-flags.mjs";
 
 // --- Episode exclusion and summary classification ---
