+++ lib/db-improvement-artifacts.mjs
@@ -1,14 +1,14 @@
 import crypto from "node:crypto";
 
 import {
-  addStringFilter,
   jsonText,
-  normalizeRepository,
-  parseJsonArray,
-  parseJsonObject,
-  validateRequiredStringField,
-} from "./data-utils.mjs";
-import { normalizeText } from "./text-normalizer.mjs";
+} from "./json-text-utils.mjs";
+import { parseJsonArray } from "./json-array-utils.mjs";
+import { parseJsonObject } from "./json-object-utils.mjs";
+import { normalizeRepository } from "./repository-utils.mjs";
+import { addStringFilter } from "./sql-filter-utils.mjs";
+import { validateRequiredStringField } from "./required-field-utils.mjs";
+import { normalizeText } from "./content-normalizer.mjs";
 import { MEMORY_SCOPE } from "./memory-scope.mjs";
 
 export const IMPROVEMENT_SOURCE_KIND = Object.freeze({
