+++ lib/db-migration-runner.mjs
@@ -3,7 +3,8 @@ import {
   classifyEpisodeDigest,
   classifySemanticMemory,
 } from "./memory-scope.mjs";
-import { parseJsonArray, parseJsonObject } from "./data-utils.mjs";
+import { parseJsonArray } from "./json-array-utils.mjs";
+import { parseJsonObject } from "./json-object-utils.mjs";
 import { SCHEMA_STATEMENTS, SCHEMA_VERSION } from "./schema.mjs";
 
 const PRIMARY_SCHEMA_VERSION_TABLE = "lore_schema_version";
