+++ lib/db-trace-samples.mjs
@@ -1,11 +1,9 @@
 import crypto from "node:crypto";
 
-import {
-  clampInteger,
-  normalizeRepository,
-  parseJsonArray,
-  parseJsonObject,
-} from "./data-utils.mjs";
+import { parseJsonArray } from "./json-array-utils.mjs";
+import { parseJsonObject } from "./json-object-utils.mjs";
+import { clampInteger } from "./numeric-utils.mjs";
+import { normalizeRepository } from "./repository-utils.mjs";
 
 function nowIso() {
   return new Date().toISOString();
