+++ lib/db.mjs
@@ -21,15 +21,13 @@ import {
   isStyleAddressingMemory,
 } from "./style-addressing.mjs";
 import { readTemporalQueryNormalizationEnabled } from "./rollout-flags.mjs";
-import {
-  addStringFilter,
-  clampInteger,
-  jsonText,
-  normalizeRepository,
-  parseJsonArray,
-  parseJsonObject,
-  validateRequiredStringField,
-} from "./data-utils.mjs";
+import { parseJsonArray } from "./json-array-utils.mjs";
+import { parseJsonObject } from "./json-object-utils.mjs";
+import { jsonText } from "./json-text-utils.mjs";
+import { clampInteger } from "./numeric-utils.mjs";
+import { normalizeRepository } from "./repository-utils.mjs";
+import { addStringFilter } from "./sql-filter-utils.mjs";
+import { validateRequiredStringField } from "./required-field-utils.mjs";
 import {
   extractFtsTerms as extractNormalizedFtsTerms,
   extractDirectTerms as extractNormalizedDirectTerms,
@@ -40,7 +38,7 @@ import {
   sanitizeFtsQuery as sanitizeNormalizedFtsQuery,
   tokenizeText,
 } from "./query-normalizer.mjs";
-import { normalizeText } from "./text-normalizer.mjs";
+import { normalizeText } from "./content-normalizer.mjs";
 import { estimateTokens } from "./token-estimator.mjs";
 import {
   appendPromptCrossRepoHintsSection,
