+++ lib/json-array-utils.mjs
@@ -1,0 +1,11 @@
+export function parseJsonArray(value) {
+  if (typeof value !== "string" || value.trim().length === 0) {
+    return [];
+  }
+  try {
+    const parsed = JSON.parse(value);
+    return Array.isArray(parsed) ? parsed : [];
+  } catch {
+    return [];
+  }
+}
