+++ lib/json-text-utils.mjs
@@ -1,0 +1,3 @@
+export function jsonText(value) {
+  return JSON.stringify(value ?? []);
+}
