+++ lib/memory-operations.mjs
@@ -13,7 +13,7 @@ import {
 } from "./workstream-overlays.mjs";
 import { buildOnboardingSection } from "./onboarding.mjs";
 import { collectFilteredReasonSummaries } from "./filtered-reason-summary.mjs";
-import { normalizeText } from "./text-normalizer.mjs";
+import { normalizeText } from "./prompt-text-normalizer.mjs";
 import { estimateTokens } from "./token-estimator.mjs";
 
 function fetchDirectives({ db, repository, includeOtherRepositories, config }) {
