+++ lib/memory-scope.mjs
@@ -1,4 +1,4 @@
-import { normalizeText } from "./text-normalizer.mjs";
+import { normalizeText } from "./content-normalizer.mjs";
 
 export const MEMORY_SCOPE = Object.freeze({
   GLOBAL: "global",
