+++ lib/memory-tools-reports.mjs
@@ -32,7 +32,9 @@ import {
   formatDoctorSafetyGateSection,
   formatReviewGateReport,
 } from "./memory-tools-observability-reports.mjs";
-import { ensureArray, ensureLimit, formatRows } from "./memory-tools-helpers.mjs";
+import { ensureArray } from "./memory-tools-array-utils.mjs";
+import { ensureLimit } from "./memory-tools-validation-utils.mjs";
+import { formatRows } from "./memory-tools-render-utils.mjs";
 
 function formatRecallSummary(trace) {
   return Object.entries(trace?.lookups ?? {})
