+++ lib/overlay-hydrator.mjs
@@ -3,7 +3,7 @@ import { existsSync } from "node:fs";
 import { readFile } from "node:fs/promises";
 import { DatabaseSync } from "node:sqlite";
 
-import { normalizeText } from "./text-normalizer.mjs";
+import { normalizeText } from "./session-text-normalizer.mjs";
 import { buildWorkstreamOverlayMemory } from "./workstream-overlays.mjs";
 import { readWorkstreamOverlaysEnabled } from "./rollout-flags.mjs";
 import { sanitizeRetainedList, sanitizeRetainedText } from "./retention-sanitizer.mjs";
