+++ lib/repository-utils.mjs
@@ -1,0 +1,5 @@
+export function normalizeRepository(repository) {
+  return typeof repository === "string" && repository.trim().length > 0
+    ? repository.trim()
+    : null;
+}
