+++ lib/required-field-utils.mjs
@@ -1,0 +1,7 @@
+export function validateRequiredStringField(value, fieldName) {
+  const normalized = String(value || "").trim();
+  if (!normalized) {
+    throw new Error(`${fieldName} is required`);
+  }
+  return normalized;
+}
