+++ lib/retention-sanitizer.mjs
@@ -14,7 +14,7 @@ const INJECTED_SECTION_HEADINGS = [
   /^## Active Workstream$/m,
 ];
 
-import { normalizeText } from "./text-normalizer.mjs";
+import { normalizeText } from "./content-normalizer.mjs";
 
 export function stripInjectedContext(value) {
   let text = String(value || "");
