+++ lib/session-store-reader.mjs
@@ -4,7 +4,7 @@ import { existsSync, readFileSync } from "node:fs";
 import { DatabaseSync } from "node:sqlite";
 import { parseSimpleYaml } from "./workspace-reader.mjs";
 import { sanitizeFtsQuery, tokenizeText } from "./query-normalizer.mjs";
-import { normalizeText } from "./text-normalizer.mjs";
+import { normalizeText } from "./session-text-normalizer.mjs";
 
 function buildSessionMatchRecord(row, hydrated, score) {
   return {
