+++ lib/trace-recorder.mjs
@@ -1,4 +1,4 @@
-import { clampInteger } from "./data-utils.mjs";
+import { clampInteger } from "./numeric-utils.mjs";
 
 function ensureArray(value) {
   return Array.isArray(value) ? value : [];
