+++ tests/unit/trace-recorder.test.mjs
@@ -15,7 +15,7 @@ async function loadTraceHotspots() {
     const tracePath = path.join(REPO_ROOT, "lib", "trace-recorder.mjs");
     const traceUrl = pathToFileURL(tracePath).href;
     const source = readFileSync(tracePath, "utf8")
-      .replace(/from "\.\/data-utils\.mjs"/g, `from "${pathToFileURL(path.join(REPO_ROOT, "lib", "data-utils.mjs")).href}"`)
+      .replace(/from "\.\/numeric-utils\.mjs"/g, `from "${pathToFileURL(path.join(REPO_ROOT, "lib", "numeric-utils.mjs")).href}"`)
       .replace("function buildTraceRecord(event, options, index) {", "export function buildTraceRecord(event, options, index) {")
       .replace("function normalizeRecorderOptions(config) {", "export function normalizeRecorderOptions(config) {");
     traceHotspotsPromise = import(`data:text/javascript;base64,${Buffer.from(`${source}\n//# sourceURL=${traceUrl}\n`).toString("base64")}`);
